// src/generated/model/vedleggFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type {VedleggRadFrontend} from "./vedleggRadFrontend";
import type {FilFrontend} from "./filFrontend";

export interface VedleggFrontend {
    type?: string;
    gruppe?: string;
    rader?: VedleggRadFrontend[];
    vedleggStatus?: string;
    filer?: FilFrontend[];
}